aac: clippy round-9 — zero warnings crate-wide
Silence clippy::excessive_precision on the included normative SBR tables
(`src/sbr/tables_data.rs` is `include!`-ed from `src/sbr/tables.rs`, so
the `#![allow]` lives on the containing module). The literal form of the
Annex 4.A coefficients is spec-pinned and preserved byte-for-byte; the
f32 values the compiler materializes are unchanged.

Two leftover test lints cleaned up narrowly:
- `tests/sbr_he_aac_psnr.rs`: drop redundant `channels as u16` cast.
- `tests/sbr_smoke.rs`: switch `SbrChannelData::default()` + field
  reassignments to struct-update syntax.

Verification:
- `cargo clippy -p oxideav-aac --all-targets -- -D warnings` exits 0.
- `cargo fmt --check -p oxideav-aac` exits 0.
- All 122 tests pass.
- HE-AACv1 stereo best-window PSNR unchanged at 48.28 dB.

+++ tests/sbr_he_aac_psnr.rs
@@ -46,7 +46,7 @@ fn write_sine_wav(path: &Path, sr: u32, freq: f32, secs: f32, channels: u16) {
         }
     }
     let byte_rate = sr * 2 * channels as u32;
-    let block_align = 2 * channels as u16;
+    let block_align = 2 * channels;
     let mut wav = Vec::with_capacity(pcm.len() + 44);
     wav.extend_from_slice(b"RIFF");
     wav.extend_from_slice(&((pcm.len() + 36) as u32).to_le_bytes());
